+++ src/frontend/tests/defaults.ts
@@ -1,0 +1,1 @@
+export const classicUrl = 'http://127.0.0.1:8000';
